requirements.txt
flask
pandas
scikit-learn
flask_cors
skforecast==0.13.0
statsmodels
prophet
google-genai
ipython
tabulate
markdown2
numpy==1.26.0